+++ src/letsbuilda/pypi/__init__.py
@@ -1,7 +1,7 @@
 """A wrapper for PyPI's API and RSS feed."""
 
+from .clients.sync_client import PyPIServices
 from .models import JSONPackageMetadata, RSSPackageMetadata
-from .sync_client import PyPIServices
 
 __all__ = [
     "PyPIServices",
