+++ src/letsbuilda/pypi/clients/async_client.py
@@ -10,7 +10,7 @@ except ImportError as error:
     msg = "Please install letsbuilda[async] for async support!"
     raise ImportError(msg) from error
 
-from .models import JSONPackageMetadata, RSSPackageMetadata
+from letsbuilda.pypi.models import JSONPackageMetadata, RSSPackageMetadata
 
 
 class PyPIServices:
