+++ src/letsbuilda/pypi/clients/sync_client.py
@@ -5,7 +5,7 @@ from typing import Final, Self
 import xmltodict
 from requests import Session
 
-from .models import JSONPackageMetadata, RSSPackageMetadata
+from letsbuilda.pypi.models import JSONPackageMetadata, RSSPackageMetadata
 
 
 class PyPIServices:
